refactor: improved status query flow
